pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pymongo>=4.15.5",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
]
